v2.6.4

+++ Commands/utils/admin/shutdown.js
@@ -5,23 +5,30 @@ module.exports = {
     description: "Shutdowns the bot from discord",
     private: true,
     execute(logger, client, message, args) {
-        if (message.author.id == 411996978583699456) {
+        const server = process.env.SERVER
+        if (message.author.id == 411996978583699456 && !args[0]) {
             logger.severe("Shutdown requested from discord...");
             client.channels.cache.get("1037141235451842701").send(`Bot shutdown requested, **Uptime**: \`${prettyMilliseconds(client.uptime)}\``);
-            message.channel.send(`**Shuting down the bot...**\n**Uptime**: \`${prettyMilliseconds(client.uptime)}\``).then(() => {
+            message.channel.send(`**Shutting down the bot...**\n**Uptime**: \`${prettyMilliseconds(client.uptime)}\``).then(() => {
                 //Destroys the client, disconnects the database and exits the process
                 client.destroy();
                 global.prisma.$disconnect();
                 process.exit(0);
             })
+            return;
         }
-        else if (message.member.permissions.has("Administrator")) {
-            //If a admin executes the command, tells it to contact iTsMaat
-            message.channel.send(`Please contact the owner of this bot to execute this command (iTsMaaT#4020 or <@411996978583699456>)`);
+
+        if (message.author.id == 411996978583699456 && (args[0] == server)) {
+            client.channels.cache.get("1037141235451842701").send(`Bot shutdown requested on \`${server}\`, **Uptime**: \`${prettyMilliseconds(client.uptime)}\``);
+            message.channel.send(`**Shutting down the bot on the \`${server}\` server...**\n**Uptime**: \`${prettyMilliseconds(client.uptime)}\``).then(() => {
+
+                client.destroy();
+                global.prisma.$disconnect();
+                process.exit(0);
+            })
+            return;
         }
-        else {
-            //Doesn't let normal users execute the command
-            message.channel.send(`nah`);
-        }
+
+        message.channel.send("Invalid server name / Not enough permissions.");
     }
 }
